modified buttons, removed sample_page
