+++ streamlit/Home.py
@@ -1,3 +1,18 @@
 import streamlit as st
 
 # TODO decide what goes in the homepage
+
+st.title("Quantum Error Correction: A Neural Network Decoder for Surface Codes")
+
+st.header("What is quantum error correction?")
+st.text("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.")
+
+st.header("How does a neural network decode quantum errors?")
+st.text("Lorem ipsum dolor sit amet consectetur adipiscing elit. Quisque faucibus ex sapien vitae pellentesque sem placerat. In id cursus mi pretium tellus duis convallis. Tempus leo eu aenean sed diam urna tempor. Pulvinar vivamus fringilla lacus nec metus bibendum egestas. Iaculis massa nisl malesuada lacinia integer nunc posuere. Ut hendrerit semper vel class aptent taciti sociosqu. Ad litora torquent per conubia nostra inceptos himenaeos.")
+st.text("Lorem ipsum dolor sit amet consectetur adipiscing elit. Quisque faucibus ex sapien vitae pellentesque sem placerat. In id cursus mi pretium tellus duis convallis. Tempus leo eu aenean sed diam urna tempor. Pulvinar vivamus fringilla lacus nec metus bibendum egestas. Iaculis massa nisl malesuada lacinia integer nunc posuere. Ut hendrerit semper vel class aptent taciti sociosqu. Ad litora torquent per conubia nostra inceptos himenaeos.")
+
+st.header("Key definitions")
+st.text("Lorem — ipsum dolor sit amet consectetur adipiscing elit")
+st.text("Quisque — faucibus ex sapien vitae pellentesque sem placerat")
+st.text("In — id cursus mi pretium tellus duis convallis")
+st.text("Basis — consectetur adipiscing elit")
